+++ .gitignore
@@ -1,2 +1,6 @@
 *.egg-info
 __pycache__
+build
+.venv
+.vscode
+site
