--- requirements.txt
@@ -1,3 +1,0 @@
-ruff==0.1.14
-isort==5.13.2
-pre-commit==3.6.0
